verificaciones devolucion ticket

+++ src/app/components/ticket-controller/list-ticket/list-ticket.component.ts
@@ -38,39 +38,26 @@ export class ListTicketComponent implements OnInit{
   {
     return this.listaTickets.length;
   }
-
-  parseDate(dateString: Date | null): number {
-    if (dateString === null) {
-      return 0; // Otra acción apropiada si la fecha es nula
-    }
   
-    // Obtén solo la fecha (sin la hora) para hacer la comparación
-    const dateOnly = new Date(
-      dateString.getFullYear(),
-      dateString.getMonth(),
-      dateString.getDate(),
-      dateString.getHours(),
-      dateString.getMinutes()
-    );
-  
-    return dateOnly.getTime();
-  }
-  
-  compareDates(itemFecha: string | null): boolean {
+  compareDates(itemFecha: string | null): boolean 
+  {  
     if (itemFecha === null) {
-      return false; // Otra acción apropiada si la fecha es nula
+      console.log("La fecha es nula.");
+      return false;
     }
-    // Convierte las cadenas de fecha a objetos Date
-    const fechaItem = new Date(itemFecha);
-    const fechaOtra = new Date(this.currentDate);
-    // Compara las fechas
-    if (fechaItem < fechaOtra) {
-        return true; // itemFecha es menor que otraFecha
+    const partesFecha = itemFecha.split('/');
+    const dia = parseInt(partesFecha[0], 10);
+    const mes = parseInt(partesFecha[1], 10) - 1; // Restamos 1 al mes porque en TS los meses van de 0 a 11
+    const año = parseInt(partesFecha[2], 10);
+  
+    const fechaProporcionada = new Date(año, mes, dia);
+    if (fechaProporcionada > this.currentDate) {
+      return true;
+    } 
+    else {
+      return false;
     }
-     else {
-        return false; // Las fechas son iguales
-    }
-}
+  }
 
   public getTickets()
   {
@@ -86,48 +73,54 @@ export class ListTicketComponent implements OnInit{
   public async DevolverTicket(ticketDevuelto: Ticket | null)
   {
     if(ticketDevuelto !== null){
-      this.showtimeService.getShowtimeById(ticketDevuelto.idFuncion!).subscribe((data: Showtime)=>{ 
-        var funcion = data;
-        funcion.entradasDisponible = funcion.entradasDisponible! + ticketDevuelto.cantidadEntradas!;
-        console.log(funcion.entradasDisponible)
-        this.showtimeService.editShowtime(ticketDevuelto.idFuncion!, funcion).subscribe(
+
+      const confirmacion = window.confirm("¿Desea devolver el ticket?");
+
+      if (confirmacion) 
+      {
+        this.showtimeService.getShowtimeById(ticketDevuelto.idFuncion!).subscribe((data: Showtime)=>{ 
+          var funcion = data;
+          funcion.entradasDisponible = funcion.entradasDisponible! + ticketDevuelto.cantidadEntradas!;
+          console.log(funcion.entradasDisponible)
+          this.showtimeService.editShowtime(ticketDevuelto.idFuncion!, funcion).subscribe(
+            {
+              next: (res) => {
+                if(res)
+                {
+                  console.log("Modificado con exito");
+                }
+                else
+                {
+                  console.log("No se pudo Modificar");
+                } 
+              },
+              error: () => console.log("No se pudo Modificar")
+            }
+          );
+        });
+        this.ticketService.deleteTicket(ticketDevuelto.id!).subscribe(
           {
             next: (res) => {
               if(res)
               {
-                console.log("Modificado con exito");
+                alert("Ticket devuelto con exito");
               }
               else
               {
-                console.log("No se pudo Modificar");
+                console.log("No se pudo Eliminar");
               } 
             },
-            error: () => console.log("No se pudo Modificar")
+            error: () => console.log("No se pudo Eliminar")
           }
         );
-      });
-      this.ticketService.deleteTicket(ticketDevuelto.id!).subscribe(
-        {
-          next: (res) => {
-            if(res)
-            {
-              console.log("Eliminado con exito");
-            }
-            else
-            {
-              console.log("No se pudo Eliminar");
-            } 
-          },
-          error: () => console.log("No se pudo Eliminar")
-        }
-      );
+      }
+      this.recargarComponente()
     }
-    this.recargarComponente()
   }
+
   recargarComponente() {
     // Obtén la ruta actual
     const currentUrl = this.router.url;
-
     // Navega a sí mismo para recargar el componente
     this.router.navigateByUrl('/', { skipLocationChange: true }).then(() => {
       this.router.navigate([currentUrl]);
